Add Gabon support with tests and country registration
This commit introduces support for the country Gabon in the application.
It includes updates to the `CountryInitializer` to register Gabon,
the implementation of the `Gabon` class adhering to the `ICountry`
interface, and the addition of unit tests in `GabonTest` to ensure
the accuracy of the country information provided by the
`CountryProvider`.